requirements.txt
openai
faiss-cpu
transformers
sentence-transformers
pinecone-client
streamlit
